Parte 2 - Aula 04
ealizar o relacionamento entre classes através de composição;
Isolar trechos de código em outros arquivos;
Importar arquivos externos em outros arquivos;
O valor do null.
