+++ parte2/alurabank/lib/cliente.dart
@@ -1,0 +1,5 @@
+class Cliente {
+  String nome;
+  String cpf;
+  String profissao;
+}
